backend/app/requirements.txt
# FastAPI项目依赖，python 3.12
fastapi==0.116.2
requests
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
httpx==0.28.1
pymysql==1.1.1
cryptography==43.0.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.1.0
typer==0.12.5
apscheduler==3.11.0
pydash==8.0.5
funcy==2.0
aiohttp==3.12.15
pyjwt==2.10.1
sqlalchemy>=2.0.0
gunicorn==21.2.0